+++ src/Database/PostgreSQL/PQTypes/JSON.hs
@@ -1,104 +1,119 @@
 module Database.PostgreSQL.PQTypes.JSON
-  ( JSON (..)
+  ( -- * Helpers, to be used with @deriving via@ (@-XDerivingVia@).
+    JSON (..)
   , JSONB (..)
-  , aesonFromSQL
-  , aesonToSQL
+
+    -- * Unparsed JSON values
+  , RawJSON (..)
+  , RawJSONB (..)
   ) where
 
-import Control.Exception qualified as E
 import Data.Aeson
 import Data.ByteString.Char8 qualified as BS
 import Data.ByteString.Lazy.Char8 qualified as BSL
+import Data.Typeable
 import Foreign.Ptr
 
 import Database.PostgreSQL.PQTypes.Format
 import Database.PostgreSQL.PQTypes.FromSQL
 import Database.PostgreSQL.PQTypes.Internal.C.Types
+import Database.PostgreSQL.PQTypes.Internal.Utils
 import Database.PostgreSQL.PQTypes.ToSQL
 
--- | Wrapper for (de)serializing underlying type as 'json'.
-newtype JSON json = JSON {unJSON :: json}
+-- | Wrapper that serializes and deserializes the underlying type as @json@
+-- with its 'ToJSON' and 'FromJSON' instances.
+--
+-- /Note:/ To get the SQL instances for a type of your own, use
+-- @DerivingVia@:
+--
+-- @
+-- data Foo = ...
+--   deriving anyclass (FromJSON, ToJSON)
+--   deriving (PQFormat, ToSQL, FromSQL) via JSON Foo
+-- @
+newtype JSON a = JSON {unJSON :: a}
   deriving (Eq, Functor, Ord, Show)
 
-instance PQFormat (JSON json) where
+instance PQFormat (JSON a) where
   pqFormat = BS.pack "%json"
 
-instance FromSQL (JSON BS.ByteString) where
-  type PQBase (JSON BS.ByteString) = PGbytea
-  fromSQL = fmap JSON . fromSQL
-
-instance FromSQL (JSON BSL.ByteString) where
-  type PQBase (JSON BSL.ByteString) = PGbytea
-  fromSQL = fmap JSON . fromSQL
-
-instance ToSQL (JSON BS.ByteString) where
-  type PQDest (JSON BS.ByteString) = PGbytea
-  toSQL = toSQL . unJSON
-
-instance ToSQL (JSON BSL.ByteString) where
-  type PQDest (JSON BSL.ByteString) = PGbytea
-  toSQL = toSQL . unJSON
-
-instance FromSQL (JSON Value) where
-  type PQBase (JSON Value) = PGbytea
+instance (FromJSON a, Typeable a) => FromSQL (JSON a) where
+  type PQBase (JSON a) = PGbytea
   fromSQL = fmap JSON . aesonFromSQL
 
-instance ToSQL (JSON Value) where
-  type PQDest (JSON Value) = PGbytea
+instance ToJSON a => ToSQL (JSON a) where
+  type PQDest (JSON a) = PGbytea
   toSQL = aesonToSQL . unJSON
 
 ----------------------------------------
 
--- | Wrapper for (de)serializing underlying type as 'jsonb'.
-newtype JSONB jsonb = JSONB {unJSONB :: jsonb}
+-- | Wrapper that serializes and deserializes the underlying type as @jsonb@
+-- with its 'ToJSON' and 'FromJSON' instances.
+--
+-- /Note:/ To get the SQL instances for a type of your own, use
+-- @DerivingVia@:
+--
+-- @
+-- data Foo = ...
+--   deriving anyclass (FromJSON, ToJSON)
+--   deriving (PQFormat, ToSQL, FromSQL) via JSONB Foo
+-- @
+newtype JSONB a = JSONB {unJSONB :: a}
   deriving (Eq, Functor, Ord, Show)
 
-instance PQFormat (JSONB jsonb) where
+instance PQFormat (JSONB a) where
   pqFormat = BS.pack "%jsonb"
 
-instance FromSQL (JSONB BS.ByteString) where
-  type PQBase (JSONB BS.ByteString) = PGbytea
-  fromSQL = fmap JSONB . fromSQL
-
-instance FromSQL (JSONB BSL.ByteString) where
-  type PQBase (JSONB BSL.ByteString) = PGbytea
-  fromSQL = fmap JSONB . fromSQL
-
-instance ToSQL (JSONB BS.ByteString) where
-  type PQDest (JSONB BS.ByteString) = PGbytea
-  toSQL = toSQL . unJSONB
-
-instance ToSQL (JSONB BSL.ByteString) where
-  type PQDest (JSONB BSL.ByteString) = PGbytea
-  toSQL = toSQL . unJSONB
-
-instance FromSQL (JSONB Value) where
-  type PQBase (JSONB Value) = PGbytea
+instance (FromJSON a, Typeable a) => FromSQL (JSONB a) where
+  type PQBase (JSONB a) = PGbytea
   fromSQL = fmap JSONB . aesonFromSQL
 
-instance ToSQL (JSONB Value) where
-  type PQDest (JSONB Value) = PGbytea
+instance ToJSON a => ToSQL (JSONB a) where
+  type PQDest (JSONB a) = PGbytea
   toSQL = aesonToSQL . unJSONB
 
 ----------------------------------------
 
--- | Helper for defining 'FromSQL' instance for a type with 'FromJSON' instance.
---
--- @since 1.9.1.0
-aesonFromSQL :: FromJSON t => Maybe PGbytea -> IO t
-aesonFromSQL mbase = do
-  evalue <- eitherDecodeStrict' <$> fromSQL mbase
-  case evalue of
-    Left err -> E.throwIO . E.ErrorCall $ "aesonFromSQL: " ++ err
-    Right value -> pure value
+-- | A @json@ value as its unparsed UTF-8 text.
+newtype RawJSON = RawJSON {unRawJSON :: BS.ByteString}
+  deriving (Eq, Ord, Show)
 
--- | Helper for defining 'ToSQL' instance for a type with 'ToJSON' instance.
---
--- @since 1.9.1.0
-aesonToSQL
-  :: ToJSON t
-  => t
-  -> ParamAllocator
-  -> (Ptr PGbytea -> IO r)
-  -> IO r
+instance PQFormat RawJSON where
+  pqFormat = BS.pack "%json"
+
+instance FromSQL RawJSON where
+  type PQBase RawJSON = PGbytea
+  fromSQL = fmap RawJSON . fromSQL
+
+instance ToSQL RawJSON where
+  type PQDest RawJSON = PGbytea
+  toSQL = toSQL . unRawJSON
+
+----------------------------------------
+
+-- | A @jsonb@ value as its unparsed UTF-8 text.
+newtype RawJSONB = RawJSONB {unRawJSONB :: BS.ByteString}
+  deriving (Eq, Ord, Show)
+
+instance PQFormat RawJSONB where
+  pqFormat = BS.pack "%jsonb"
+
+instance FromSQL RawJSONB where
+  type PQBase RawJSONB = PGbytea
+  fromSQL = fmap RawJSONB . fromSQL
+
+instance ToSQL RawJSONB where
+  type PQDest RawJSONB = PGbytea
+  toSQL = toSQL . unRawJSONB
+
+----------------------------------------
+
+aesonFromSQL :: forall a. (FromJSON a, Typeable a) => Maybe PGbytea -> IO a
+aesonFromSQL mbase = do
+  v <- fromSQL mbase
+  case eitherDecodeStrict' v of
+    Right a -> pure a
+    Left err -> hpqTypesError $ "aesonFromSQL (" ++ show (typeRep $ Proxy @a) ++ "): " ++ err
+
+aesonToSQL :: ToJSON a => a -> ParamAllocator -> (Ptr PGbytea -> IO r) -> IO r
 aesonToSQL = toSQL . BSL.toStrict . encode